requirements.txt
anthropic
httpx
playwright
beautifulsoup4
lxml
sqlite-utils
python-dotenv
